+++ tecte/Programmers/PROG_0000_template.java
@@ -1,0 +1,21 @@
+/*
+ * 문제 : 프로그래머스 - 0000 (https://school.programmers.co.kr/learn/courses/30/lessons/0000)
+ * 문제 풀이
+ *
+ */
+
+package Programmers;
+
+public class PROG_0000_template {
+    static class Solution {
+        public int solution() {
+
+
+            return 0;
+        }
+    }
+
+    public static void main(String[] args) {
+
+    }
+}
